+++ crates/forge-runner/src/forge_runtime_extension/file_operations.rs
@@ -94,7 +94,7 @@ pub(super) fn string_into_felt(string: &str) -> Result<Felt252> {
 
 #[cfg(test)]
 mod tests {
-    use crate::test_execution_syscall_handler::file_operations::string_into_felt;
+    use crate::forge_runtime_extension::file_operations::string_into_felt;
     use cairo_felt::Felt252;
     use num_bigint::BigUint;
     use num_traits::One;
